Move validateTableLength to its own file
