+++ script/fetch-recorded-data
@@ -42,6 +42,7 @@ require('ts-node').register({
   }
 })
 const internationalParser = require('../lib/international-interventions-parser')
+const validateFetch = require('../lib/validate-fetch')
 
 const cacheDir = argv.cacheDir
 if (cacheDir) {
@@ -266,7 +267,7 @@ async function main() {
       [caseRecords]
     )
 
-    await internationalParser.validateTableLength(
+    await validateFetch.validateTableLength(
       db,
       'case_data',
       'case_data_import',
@@ -314,7 +315,7 @@ async function main() {
       [allInterventionRecords]
     )
 
-    await internationalParser.validateTableLength(
+    await validateFetch.validateTableLength(
       db,
       'intervention_data',
       'intervention_data_import',
